# optimize route building when there's query params.