--- a/shard.yml
+++ b/shard.yml
@@ -1,7 +1,7 @@
 name: lucky
 version: 1.1.0
 
-crystal: ">=1.6.0"
+crystal: ">=1.9.0"
 
 authors:
   - Paul Smith <paulcsmith0218@gmail.com>